Done!
The compiled part is (mostly) done, anyway. What remains is to actually write newplot.C, a macro that will take trees from trackeffic.root and fill (mostly 3D) histograms with the leaves which will then be projected out into 1D or 2D histograms for slices of interest. I also will keep brainstorming with ways to pull more things from the main function to encapsulate processes in the namespace. So consider trackeffic.cxx a rough draft of the overhaul.

What's new: changed the Makefile completely. Now run simply with "make".
Wrote analysis_funcs, .cxx and .h. These contain a custom namespace used in trackeffic.cxx, and within the namespace there's a class for creating/filling containers - i.e. the trees, the branches, and the particle vectors.
uncutClass, .C and .h are very, very, very, very preliminary attempts to read the trees and output histograms. (So far it works for one branch of one tree! Just remains to be generalized to all branches/trees, and once done for each individually, encoded into a